requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
bitarray==3.6.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ckzg==2.1.1
click==8.2.1
cryptography==45.0.6
cyclopts==3.22.5
cytoolz==1.0.1
deprecation==2.1.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.3.0
fastmcp==2.11.2
frozenlist==1.7.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.4
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.6.3
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==25.0
parse==1.20.2
parsimonious==0.10.0
pathable==0.4.4
postgrest==1.1.1
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyunormalize==16.0.0
PyYAML==6.0.2
realtime==2.7.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rlp==4.1.0
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.0
supafunc==0.10.1
toolz==1.0.0
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
web3==7.13.0
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.20.1
